Clean up the bootload logic
